+++ modules/spring-web/src/main/kotlin/com/esprito/spring/web/SpringWebIcons.kt
@@ -1,0 +1,9 @@
+package com.esprito.spring.web
+
+import com.intellij.openapi.util.IconLoader
+
+object SpringWebIcons {
+    private fun load(path: String) = IconLoader.getIcon(path, SpringWebIcons.javaClass)
+
+    val OpenApi = load("com/esprito/spring/web/icons/openApi.svg")
+}